Issue #114 handle `ctags -n` jumping to symbol
symbols-view expects there to be patterns in `tags` files. This is the default behavior when running ctags (you can also explictly set it with the -F, or --excmd=pattern options). When running ctags with the number option (-n or --excmd=number), e.g ctags -R -n, numbers are used instead of patterns to identify lines.

This package should expect and accept both formats. To determine which format a tag uses, check if tag.pattern and tag.lineNumber represent the same number.  If so create a Point based on line number instead of pattern.  `node-ctags` populates the tag.pattern parameter with the stringified lineNumber if `ctags -n` is run. When a tag has a pattern instead of a lineNumber,  tag.LineNumber defaults to 0, and pattern will not be compares unless it is a line containing only a number, so will almost always return False.

An additional test fixture was added for tags generated using the `ctags -R -n`.