Allow end-user configuration of tags path relative to project root
Allow user to specify the directory, relative to the root of each project
open within Atom, where symbols-view looks for the `tags` file, using
the `tagsDirectory` config variable.

This only effects `toggle-project-symbols`. The impetus behind this commit
is to allow popular confirations, such as storing tags files in the `.git`
directory, to work. The change required additional information on each
`tag` object, mainly specifiying the path of the project the tag was generated
for. This was to a) easily display the project directory when multiple projects
were open and b) to programatically resolve the full path of the symbol's file
when generating symbols relative to root.

This also required adding a configuration variable `tagsDirectory`, which should
be a directory relative to the project root that contains the tags file.

Also refactored the reloadTags method inside of watchTagsFiles into a separate
instance method, with a different name (triggerReloadTags) so that it didn't
shadow the instance attribute reloadTags, and could be used in the configuration
watcher for the new `tagsDirectory` configuration variable.

Added an appropriate test to run `toggle-project-symbols` with a relative
`tagsDirectory` config set.