Fix getting word under cursor for various Ruby scopes
In Ruby blocks within ERB, go-to-declaration wouldn't work when cursor
was positioned on identifiers such as `foo?` or `foo!` because the
`?!` characters wouldn't be considered part of the current "word". This
is due to the fact that the previous special case for `source.ruby`
scope didn't apply for ERB where the scope is `source.ruby.embedded.erb`.