Enhance handling of fully-qualified Ruby constants
When cursor is over word `Foo` in `A::Foo::B`, don't look up `A::Foo::B`
constant first because that might not be what the user wants. Instead,
look up `A::Foo` with a fall back to `Foo`. By looking only the fully
qualified constant portion up until the cursor position, we avoid the
risk of throwing the user in an unwanted direction.