Added symbols-view.extraCtagsArguments
This allows to add extra arguments to the ctags invocation
used by file-view, an example use case is using --options and supplying
an extra ctags config file which should only be used with Atom. That is
the user doesn't want to place it in $HOME/.ctags or a similar location.